+++ src/components/PopupResult/PopupResult.jsx
@@ -2,29 +2,29 @@ import React, { useEffect, useState } from 'react'
 import PropTypes from 'prop-types'
 import './PopupResult.css'
 import { useSelector } from 'react-redux'
-import { debounceTitilerOverlay, zoomToItemExtent } from '../../utils/mapHelper'
+import { zoomToItemExtent } from '../../utils/mapHelper'
 import ItemHeader from '../EnhancedDetails/ItemHeader.jsx'
+import VisualizationDropdown from '../VisualizationDropdown/VisualizationDropdown'
 
 const PopupResult = (props) => {
   const _appConfig = useSelector((state) => state.mainSlice.appConfig)
   const _autoCenterOnItemChanged = useSelector(
     (state) => state.mainSlice.autoCenterOnItemChanged
   )
-  const [thumbnailURL, setthumbnailURL] = useState(null)
+  const [thumbnailInfo, setThumbnailInfo] = useState(null)
 
   useEffect(() => {
     if (props.result) {
       if (_autoCenterOnItemChanged) {
         zoomToItemExtent(props.result)
       }
-      debounceTitilerOverlay(props.result)
       const thumbnailURLForSelection = props.result?.links?.find(
         ({ rel }) => rel === 'thumbnail'
       )?.href
 
       // If no thumbnail available, clear immediately
       if (!thumbnailURLForSelection) {
-        setthumbnailURL(null)
+        setThumbnailInfo(null)
         return
       }
 
@@ -32,7 +32,11 @@ const PopupResult = (props) => {
       const image = new Image()
       image.onload = function () {
         if (this.width > 0) {
-          setthumbnailURL(thumbnailURLForSelection)
+          setThumbnailInfo({
+            url: thumbnailURLForSelection,
+            width: this.width,
+            height: this.height
+          })
         }
       }
       image.src = thumbnailURLForSelection
@@ -51,11 +55,16 @@ const PopupResult = (props) => {
     >
       {props.result ? (
         <div className="popupResultHero">
-          {thumbnailURL && (
-            <div className="popupResultThumbnailContainer">
+          {thumbnailInfo && (
+            <div
+              className="popupResultThumbnailContainer"
+              style={{
+                aspectRatio: `${thumbnailInfo.width} / ${thumbnailInfo.height}`
+              }}
+            >
               <picture>
                 <img
-                  src={thumbnailURL}
+                  src={thumbnailInfo.url}
                   alt="thumbnail"
                   className="popupResultThumbnail"
                   onError={({ currentTarget }) => {
@@ -66,10 +75,9 @@ const PopupResult = (props) => {
               </picture>
             </div>
           )}
-          <ItemHeader
-            id={props.result.id}
-            collection={props.result.collection}
-          />
+          <ItemHeader id={props.result.id} collection={props.result.collection}>
+            <VisualizationDropdown />
+          </ItemHeader>
         </div>
       ) : null}
     </div>